requirements.txt
pandas==2.2.1
fastapi==0.110.0
bitsandbytes==0.42.0
peft==0.8.2
trl==0.7.10
accelerate==0.27.1
datasets==2.17.0
transformers==4.38.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
docstring-parser==0.15
exceptiongroup==1.2.0

filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
h11==0.14.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
packaging==23.2

psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
safetensors==0.4.2
scipy==1.12.0
shtab==1.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
typing_extensions==4.9.0
tyro==0.7.3
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
xxhash==3.4.1
yarl==1.9.4
